+++ packages/react-fabric-taxonomypicker/src/api/TaxonomyApi/TaxonomyApi.ts
@@ -10,9 +10,7 @@ export interface ITaxonomyApiContext {
 }
 
 export class TaxonomyApi {
-  private static CACHEDATA: {
-    [x: string]: ITerm[];
-  } = {};
+  private static CACHEDATA: { [x: string]: ITerm[] } = {};
 
   private spContext: SP.ClientContext;
   private cacheKey: string;
@@ -77,7 +75,11 @@ export class TaxonomyApi {
     return matchingTerms;
   }
 
-  public async getTermTree(): Promise<{ termSetName: string; isOpenTermSet: boolean; terms: ITerm[] }> {
+  public async getTermTree(): Promise<{
+    termSetName: string;
+    isOpenTermSet: boolean;
+    terms: ITerm[];
+  }> {
     const taxonomySession = SP.Taxonomy.TaxonomySession.getTaxonomySession(this.spContext);
     const termStore = taxonomySession.getDefaultSiteCollectionTermStore();
     const termSet = termStore.getTermSet(new SP.Guid(this.context.termSetId));
@@ -93,6 +95,7 @@ export class TaxonomyApi {
       path: term.path,
       isSelectable: term.properties!.isSelectable || false,
       parentId: term.properties!.parentId || null,
+      sortOrder: term.sortOrder,
       children: []
     }));
 
@@ -144,11 +147,13 @@ export class TaxonomyApi {
       name: newTerm.get_name(),
       path: newTerm.get_pathOfTerm(),
       properties: {
-          isSelectable: true,
-          parentId: !parentTerm
-            ? (!!term ? this.context.rootTermId : undefined)
-            : parentTerm.get_id().toString()
-        }
+        isSelectable: true,
+        parentId: !parentTerm
+          ? !!term
+            ? this.context.rootTermId
+            : undefined
+          : parentTerm.get_id().toString()
+      }
     };
 
     // Update the cache in the correct node
@@ -191,7 +196,7 @@ export class TaxonomyApi {
 
     this.spContext.load(termSet);
     this.spContext.load(matchingTerms);
-    this.spContext.load(matchingTerms, "Include(Labels, Parent, Parent.Id)");
+    this.spContext.load(matchingTerms, "Include(Labels, Parent, Parent.Id, CustomSortOrder)");
     await this.awaitableExecuteQuery(this.spContext);
 
     const terms: ITerm[] = [];
@@ -224,6 +229,7 @@ export class TaxonomyApi {
 
       terms.push({
         id: currentTerm.get_id().toString(),
+        sortOrder: currentTerm.get_customSortOrder(),
         name: currentTerm.get_name(),
         labels: labels,
         path: currentTerm.get_pathOfTerm(),
@@ -283,6 +289,58 @@ export class TaxonomyApi {
         }
       }
     }
+
+    // Next iteration -> sort the entire tree
+    return this._sortTree(tree);
+  }
+
+  private _sortTree(tree: ITermData[]) {
+    for (let i = 0; i < tree.length; i++) {
+      tree[i] = this._sortTermAndChildren(tree[i]);
+    }
     return tree;
+  }
+
+  private _sortTermAndChildren(term: ITermData): ITermData {
+    if (term.children.length && term.sortOrder) {
+      // If not null, the custom sort order is a string of GUIDs, delimited by a :
+      if (term.sortOrder) {
+        const sortOrderSplit = term.sortOrder.split(":");
+
+        term.children.sort((a, b) => {
+          if (!!a.id && !!b.id) {
+            const indexA = sortOrderSplit.indexOf(a.id);
+            const indexB = sortOrderSplit.indexOf(b.id);
+
+            if (indexA > indexB) {
+              return 1;
+            } else if (indexA < indexB) {
+              return -1;
+            }
+
+            return 0;
+          }
+
+          return 0;
+        });
+      } else {
+        // If null, terms are just sorted alphabetically
+        term.children.sort((a, b) => {
+          if (a.name > b.name) {
+            return 1;
+          } else if (a.name < b.name) {
+            return -1;
+          }
+
+          return 0;
+        });
+      }
+    }
+
+    for (let i = 0; i < term.children.length; i++) {
+      term.children[i] = this._sortTermAndChildren(term.children[i]);
+    }
+
+    return term;
   }
 }